add Timber

+++ app/src/main/java/com/ozgursakizli/noteapplication/database/notes/NoteEntity.kt
@@ -7,17 +7,13 @@ import com.ozgursakizli.noteapplication.database.DatabaseConstants
 
 @Entity(tableName = DatabaseConstants.NOTES_TABLE)
 data class NoteEntity(
-    @ColumnInfo(name = "title")
-    var title: String,
-    @ColumnInfo(name = "description")
-    var description: String,
-    @ColumnInfo(name = "imageUrl")
-    var imageUrl: String?,
-    @ColumnInfo(name = "createdDate")
-    val createdDate: Long
+    @ColumnInfo(name = "title") var title: String,
+    @ColumnInfo(name = "description") var description: String,
+    @ColumnInfo(name = "imageUrl") var imageUrl: String?,
+    @ColumnInfo(name = "createdDate") val createdDate: Long
 ) {
     @PrimaryKey(autoGenerate = true)
-    var id: Long = 0
+    var id: Int = 0
 
     override fun toString(): String {
         return "id: $id, title: $title, description: $description, imageUrl: $imageUrl, createdDate: $createdDate"
