+++ app/src/main/java/com/ozgursakizli/noteapplication/database/notes/NotesDataSource.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface NotesDataSource {
     fun getAllNotes(): Flow<List<NoteEntity>>
-    fun getNoteById(noteId: Long): Flow<NoteEntity>
+    fun getNoteById(noteId: Int): Flow<NoteEntity>
     suspend fun insert(noteEntity: NoteEntity): Long
     suspend fun delete(noteEntity: NoteEntity)
 }
